requirements.txt
pillow
Spire.Doc
git+https://github.com/RedFantom/ttkthemes
pdf2image